Fit sinusoids with frequency defined by a polynomial.
Supports polynomial definition of amplitude everywhere but fitting.
Very hacky, but works on synthetic data with variable frequency
and constant phase. The observed data is not fitted well, which
probably indicates that the phase is unstable (which is obvious
in hindsight).

Next step will be modifying animated_window to track the phase
behavior, and include this in the fit.